add burnable to examples since function macro export was invalid
