+++ Cargo.toml
@@ -1,13 +1,14 @@
 [workspace]
 members = [
-  "contracts",
-  "lib/crypto",
-  "lib/grip",
-  "lib/grip-proc",
-  "examples/erc20",
-  "examples/erc721",
-  "examples/merkle-proofs",
-  "examples/ownable",
+    "contracts",
+    "lib/crypto",
+    "lib/grip",
+    "lib/grip-proc",
+    "examples/erc20",
+    "examples/erc721",
+    "examples/merkle-proofs",
+    "examples/ownable",
+    "examples/erc20-burnable",
 ]
 # Explicitly set the resolver to version 2, which is the default for packages
 # with edition >= 2021.
