+++ contracts/Cargo.toml
@@ -28,7 +28,6 @@ default = []
 # ERC-20
 erc20 = []
 erc20_metadata = ["erc20"]
-erc20_burnable = ["erc20"]
 
 # ERC-721
 erc721 = []
