+++ contracts/src/erc20/extensions/mod.rs
@@ -6,8 +6,3 @@ cfg_if::cfg_if! {
         pub use metadata::ERC20Metadata;
     }
 }
-cfg_if::cfg_if! {
-    if #[cfg(any(test, feature = "erc20_burnable"))] {
-        pub mod burnable;
-    }
-}
