+++ examples/erc20-burnable/Cargo.toml
@@ -1,5 +1,5 @@
 [package]
-name = "erc20-example"
+name = "erc20-burnable-example"
 edition.workspace = true
 license.workspace = true
 repository.workspace = true
@@ -7,7 +7,7 @@ publish = false
 version = "0.0.0"
 
 [dependencies]
-contracts = { path = "../../contracts", features = ["erc20_metadata"] }
+contracts = { path = "../../contracts", features = ["erc20"] }
 alloy-primitives.workspace = true
 stylus-sdk.workspace = true
 stylus-proc.workspace = true
