+++ examples/erc20/src/lib.rs
@@ -3,10 +3,7 @@ extern crate alloc;
 
 use alloc::string::String;
 
-use contracts::{
-    erc20::{extensions::ERC20Metadata, ERC20},
-    erc20_burnable_impl,
-};
+use contracts::erc20::{extensions::ERC20Metadata, ERC20};
 use stylus_sdk::prelude::{entrypoint, external, sol_storage};
 
 const DECIMALS: u8 = 10;
@@ -24,10 +21,6 @@ sol_storage! {
 #[external]
 #[inherit(ERC20, ERC20Metadata)]
 impl Token {
-    // This macro implements ERC20Burnable functions -- `burn` and `burn_from`.
-    // Expects an `ERC20 erc20` as a field of `Token`.
-    erc20_burnable_impl!();
-
     pub fn constructor(&mut self, name: String, symbol: String) {
         self.metadata.constructor(name, symbol);
     }
